requirements.txt
Django==4.1.4
environs[django]==9.3.0
Pillow==9.0.*
django-debug-toolbar==3.7.0
django-smart-selects==1.6.0
django-phonenumber-field[phonenumbers]==7.0.1